requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Pydantic & Settings
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# LLM Providers
openai==1.57.4
anthropic==0.40.0

# Memory & Vector DB
supermemory

# Document Processing
PyMuPDF==1.24.14          # PDF processing
Pillow==11.0.0             # Image processing
python-pptx==1.0.2         # PowerPoint processing

# AWS
boto3==1.35.85

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.2
google-auth==2.37.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1

# Testing & Eval
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Frontend
streamlit==1.41.1
requests==2.32.3
